packages/capl-symbol-db/pyproject.toml
[project]
name = "capl-symbol-db"
version = "0.1.0"
description = "Symbol extraction & SQLite database for CAPL"
readme = "README.md"
authors = [
    { name = "MohamedHamed19m", email = "mohamedahamed1915@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "capl-tree-sitter",
]

[tool.uv.sources]
capl-tree-sitter = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/capl_symbol_db"]